fix(personal-assistant): Date.UTC years 0-99 must use setUTCFullYear across time and analyzer helpers
Use lightweight utcDateMs util so regression test imports production without @noble/hashes heavy import. Fix shared import to lightweight subpath.
